fix(apiclient): fixed empty ConvertIDN request
function __autoIDNConvert in class hexonet.apiconnector.apiclient.APIClient
would send an empty ConvertIDN request, if cmd contains at least one
key that matches r'^(DOMAIN|NAMESERVER|DNSZONE)([0-9]*)$', but all
values at such keys match r'^[a-z0-9.-]+$'

the result of an empty ConvertIDN request would never result in a change
of cmd and thus can be avoided.

this commit adds a check `if not toconvert` just before sending the
(empty) request in order to avoid it. for readability, the now redundant
check `if not key.count` is removed.